try again

+++ lab1_IO.c
@@ -115,7 +115,7 @@ int Lab1_saveoutput(int **C, int *n, double Time)
 void *threadCalc(void *rank) {
   printf("thread calc commences\n");
 
-  int ranknum = (int) rank;
+  long ranknum = (long) rank;
 printf("rank is %d\n", ranknum);
   int x = floor(ranknum / ((int) sqrt((double) p)));
   int y = ranknum % ((int) sqrt((double) p));
